+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/engelsystem/Dependencies.kt
@@ -22,7 +22,7 @@ object Libs {
     private object Versions {
         const val junitJupiter = "5.9.1"
         const val kotlinCoroutines = "1.6.4"
-        const val moshi = "1.13.0"
+        const val moshi = "1.14.0"
         const val okhttp = "3.12.13"
         const val retrofit = "2.6.4"
         const val threetenbp = "1.6.4"